feat: 시세조회 관련 타입정의
