+++ src/types/market-price-type/similarProduct.types.d.ts
@@ -1,0 +1,15 @@
+import { CommonResponse } from './common.types';
+
+export type SimilarProduct = {
+  productId?: number;
+  title?: string;
+  price?: number;
+  tfidfVector?: string;
+  imageUrl?: string;
+};
+
+export type SimilarProductListResponse = CommonResponse<SimilarProduct[]>;
+
+export type SimilarProductProps = {
+  products: SimilarProduct[];
+};
